fix for setting input and output folder separately
