+++ sar_pre_processing/__init__.py
@@ -1,3 +1,4 @@
+from .attribute_dict import AttributeDict
 from .sar_pre_processor import SARPreProcessor
 from .file_list_sar_pre_processing import SARList
 from .netcdf_stack import NetcdfStack
