+++ tests/test_data/config_with_no_input_and_output_information.yml
@@ -16,12 +16,7 @@ General:
     output_directory_root: /some/where/
 
 SAR:
-  input_folder: ./tests/test_data/inputs
-  output_folder: ./tests/test_data/outputs
   gpt: \any\path\gpt  # Location of SNAP's graph-processing-tool
-  xml_graph_path: ./tests/test_data
-  xml_graph_pre_process_step1: dummy_graph_file.xml
-  xml_graph_pre_process_step1_border: ./tests/test_data_2/another_dummy_graph_file.xml
   speckle_filter:
     multi_temporal:
       apply: 'yes'
